requirements/base.txt
Django==1.7.7
django-bootstrap3==5.1.1
django-allauth==0.19.1
django-datetime-widget==0.9.3
django-braces==1.4.0
djangorestframework==3.1.1
django-angular==0.7.12
